+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/electric/EnergyRegulator.java
@@ -50,6 +50,7 @@ public class EnergyRegulator extends SlimefunItem {
     @Override
     public void preRegister() {
         addItemHandler(onPlace());
+
         addItemHandler(new BlockTicker() {
 
             @Override
@@ -65,7 +66,8 @@ public class EnergyRegulator extends SlimefunItem {
     }
 
     private void tick(Block b) {
-        EnergyNet.getNetworkFromLocationOrCreate(b.getLocation()).tick(b);
+        EnergyNet network = EnergyNet.getNetworkFromLocationOrCreate(b.getLocation());
+        network.tick(b);
     }
 
 }