+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/gps/GPSTransmitter.java
@@ -18,7 +18,6 @@ import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
 import me.mrCookieSlime.Slimefun.Objects.handlers.BlockTicker;
 import me.mrCookieSlime.Slimefun.api.BlockStorage;
 import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
-import me.mrCookieSlime.Slimefun.api.energy.ChargableBlock;
 
 public abstract class GPSTransmitter extends SimpleSlimefunItem<BlockTicker> implements EnergyNetComponent {
 
@@ -53,12 +52,12 @@ public abstract class GPSTransmitter extends SimpleSlimefunItem<BlockTicker> imp
 
             @Override
             public void tick(Block b, SlimefunItem item, Config data) {
-                int charge = ChargableBlock.getCharge(b);
+                int charge = getCharge(b.getLocation());
                 UUID owner = UUID.fromString(BlockStorage.getLocationInfo(b.getLocation(), "owner"));
 
                 if (charge >= getEnergyConsumption()) {
                     SlimefunPlugin.getGPSNetwork().updateTransmitter(b.getLocation(), owner, true);
-                    ChargableBlock.setCharge(b.getLocation(), charge - getEnergyConsumption());
+                    removeCharge(b.getLocation(), getEnergyConsumption());
                 }
                 else {
                     SlimefunPlugin.getGPSNetwork().updateTransmitter(b.getLocation(), owner, false);