+++ CHANGELOG.md
@@ -51,6 +51,8 @@
 * Small performance improvements for Reactors
 * Electric machines now show their tier in the Inventory name too
 * Removed "Fuel efficiency" attribute for androids, since that was pretty much always at 1.0 anyway...
+* Performance improvements for energy networks
+* (API) Rewritten Block-Energy API
 
 #### Fixes
 * Fixed Programmable Androids rotating in the wrong direction